+++ script_test.go
@@ -1286,39 +1286,6 @@ func TestFindFiles_InNonexistentPathReturnsError(t *testing.T) {
 	}
 }
 
-func TestFindFiles_DoesNotErrorWhenSubDirectoryIsNotReadable(t *testing.T) {
-	t.Parallel()
-	tmpDir := t.TempDir()
-	fileAPath := filepath.Join(tmpDir, "file_a.txt")
-	if err := os.WriteFile(fileAPath, []byte("hello world!"), os.ModePerm); err != nil {
-		t.Fatal(err)
-	}
-	restrictedDirPath := filepath.Join(tmpDir, "restricted_dir")
-	if err := os.Mkdir(restrictedDirPath, os.ModePerm); err != nil {
-		t.Fatal(err)
-	}
-	fileBPath := filepath.Join(restrictedDirPath, "file_b.txt")
-	if err := os.WriteFile(fileBPath, []byte("hello again!"), os.ModePerm); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.Chmod(restrictedDirPath, 0o000); err != nil {
-		t.Fatal(err)
-	}
-	t.Cleanup(func() { os.Chmod(restrictedDirPath, os.ModePerm) })
-	p := script.FindFiles(tmpDir)
-	if p.Error() != nil {
-		t.Fatal(p.Error())
-	}
-	want := fileAPath + "\n"
-	got, err := p.String()
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !cmp.Equal(want, got) {
-		t.Fatal(cmp.Diff(want, got))
-	}
-}
-
 func TestIfExists_ProducesErrorPlusNoOutputForNonexistentFile(t *testing.T) {
 	t.Parallel()
 	want := ""
